message encoder에서 header 옵션을 추가할 때, utf8로 인코딩된 bytes의 길이가 아닌 문자열의 길이를 저장하도록 함. 이를 수정하고, 인코딩 시 문자열이 길어질 수 있으므로 writeByte()에서 writeShort()로 상승 적용

+++ app/src/main/java/capstone/design/message/MessageDecoder.java
@@ -102,7 +102,7 @@ public class MessageDecoder extends ByteToMessageDecoder {
         // 헤더 읽기
         byte headerCount = in.readByte();
         for (int i = 0; i < headerCount; i++) {
-            byte keyLength = in.readByte();
+            short keyLength = in.readShort();
             String key = in.readString(keyLength, StandardCharsets.UTF_8);
 
             int valueLength = in.readInt();
